requirements.txt
openai
pydantic
python-dotenv
tqdm